Site Settings: add HoC for redirection to /settings/general
Currently the redirectToGeneral() f-n is used in ManageConnection
and soon in the DisconnectFlow components. This PR abstract
this acion into a separate HoC for reusability.

This PR creates a HoC and subsequently uses it in
the Manage Connection class.
