+++ client/my-sites/site-settings/manage-connection/index.jsx
@@ -2,9 +2,8 @@
  * External dependencies
  */
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
+import { flowRight } from 'lodash';
 import { localize } from 'i18n-calypso';
-import page from 'page';
 
 /**
  * Internal dependencies
@@ -15,39 +14,22 @@ import DocumentHead from 'components/data/document-head';
 import HeaderCake from 'components/header-cake';
 import Main from 'components/main';
 import SiteOwnership from './site-ownership';
-import { getSelectedSiteId, getSelectedSiteSlug } from 'state/ui/selectors';
-import { isJetpackSite } from 'state/sites/selectors';
-import { isSiteAutomatedTransfer } from 'state/selectors';
+import redirectNonJetpackToGeneral from
+	'my-sites/site-settings/redirect-to-general';
 
 class ManageConnection extends Component {
-	componentDidMount() {
-		this.verifySiteIsJetpack();
-	}
-
-	componentDidUpdate() {
-		this.verifySiteIsJetpack();
-	}
-
-	verifySiteIsJetpack() {
-		if ( this.props.siteIsJetpack === false || this.props.siteIsAtomic ) {
-			this.redirectToGeneral();
-		}
-	}
-
-	redirectToGeneral = () => {
-		const { siteSlug } = this.props;
-
-		page( '/settings/general/' + siteSlug );
-	};
 
 	render() {
-		const { translate } = this.props;
+		const {
+			translate,
+			redirectToGeneral
+		} = this.props;
 
 		return (
 			<Main className="manage-connection site-settings">
 				<DocumentHead title={ translate( 'Site Settings' ) } />
 
-				<HeaderCake onClick={ this.redirectToGeneral }>
+				<HeaderCake onClick={ redirectToGeneral }>
 					{ translate( 'Manage Connection' ) }
 				</HeaderCake>
 
@@ -59,14 +41,7 @@ class ManageConnection extends Component {
 	}
 }
 
-export default connect(
-	( state ) => {
-		const siteId = getSelectedSiteId( state );
-
-		return {
-			siteIsAtomic: isSiteAutomatedTransfer( state, siteId ),
-			siteIsJetpack: isJetpackSite( state, siteId ),
-			siteSlug: getSelectedSiteSlug( state ),
-		};
-	}
-)( localize( ManageConnection ) );
+export default flowRight(
+	localize,
+	redirectNonJetpackToGeneral
+)( ManageConnection );
